edit message works again
